+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.4.0
+* Add ReOrderableListWidget and ReOrderableListDockWidget
+* Add example of using the ReOrderableListWidget
+
 ## v0.3.0
 * Add ProgressTimerDialog and example.
 * Delete ErrorObserver, as this is relevant to VTK, not Qt, so it has been moved to the CILViewer repo.